Mega-Commits: Forgotten?
Nothing is new today. Some of package maintainer does not seems to
notice https://github.com/termux/termux-packages/blob/master/CONTRIBUTING.md#packaging-policy yet, especially the stated line:

> Software available in major Linux distributions has more chances to be
  included into Termux repositories.

You see, Some of packages were submitted, or added into termux-packages were mostly a packages that were not exist in some of distribution/BSDs package managers.

Not forget to mention that most patchs didn't use @TERMUX_PREFIX@ or @TERMUX_HOME@ as it seems like somebody didn't notice the problem for some forks.

All of these files were patched automatically, some were done with some issues or more. If you find there's a issue, Submit a change or review it as fast as possible.

This commit is actually squashed after some of mistakes that were done, such as in-place editing, typos, and so on. Thanks to @xtkoba for checking & reviewing the code so i could fix the mistake that were done by automated patching script.

That's it. Have a nice day

Signed-off-by: Yonle <yonle@duck.com>